+++ nodes/Crawlbrulee/descriptions/index.ts
@@ -1,8 +1,7 @@
 import type { INodeProperties } from 'n8n-workflow';
 import { accountOperations } from './account';
-import { jobFields, jobOperations } from './job';
-import { pageFields, pageOperations } from './page';
-import { siteFields, siteOperations } from './site';
+import { mapFields, mapOperations } from './map';
+import { scrapeFields, scrapeOperations } from './scrape';
 
 export const resourceProperty: INodeProperties = {
 	displayName: 'Resource',
@@ -11,20 +10,17 @@ export const resourceProperty: INodeProperties = {
 	noDataExpression: true,
 	options: [
 		{ name: 'Account', value: 'account' },
-		{ name: 'Job', value: 'job' },
-		{ name: 'Page', value: 'page' },
-		{ name: 'Site', value: 'site' },
+		{ name: 'Map', value: 'map' },
+		{ name: 'Scrape', value: 'scrape' },
 	],
-	default: 'page',
+	default: 'scrape',
 };
 
 export const crawlbruleeProperties: INodeProperties[] = [
 	resourceProperty,
-	pageOperations,
-	jobOperations,
-	siteOperations,
+	scrapeOperations,
+	mapOperations,
 	accountOperations,
-	...pageFields,
-	...jobFields,
-	...siteFields,
+	...scrapeFields,
+	...mapFields,
 ];